REFACTOR risky - move function to new file
